+++ sound-mind-studio/tests/test_grid_panel.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include <QObject>
+
+class GridPanelTest : public QObject {
+    Q_OBJECT
+
+private slots:
+    void freshPanelHasBothAxisLabelsOff();
+    void changingTheVerticalAxisComboEmitsVerticalAxisLabelModeChanged();
+    void changingTheHorizontalAxisComboEmitsHorizontalAxisLabelModeChanged();
+};